+++ CHANGELOG.md
@@ -11,6 +11,10 @@ all APIs might be changed.
 
 ## Unreleased - yyyy-mm-dd
 
+### New Features
+
+- Added and_then decoder function
+
 ## v0.3.0 - 2020-05-24
 
 ### New Features